pyproject.toml
[project]
name = "ai-study-companion"
version = "0.1.0"
description = "Container-free AI Study Companion backend and Streamlit prototype powered by Supabase."
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi~=0.115.5",
    "langchain-core~=1.0.0",
    "langchain-groq>=1,<2",
    "langgraph>=1,<2",
    "langgraph-checkpoint-postgres>=3,<4",
    "psycopg[binary]>=3.2,<4",
    "pydantic~=2.10.1",
    "pypdf~=5.3.0",
    "python-dotenv~=1.0.1",
    "sentence-transformers>=3.3.1",
    "streamlit~=1.52.0",
    "supabase>=2.10.0",
    "uvicorn~=0.32.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "U"]
